requirements.txt
qq-botpy>=1.2.1
PyYAML>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
